+++ .gitignore
@@ -1,2 +1,3 @@
 # remove the large data file
-data
+node_modules
+db
