+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "sv-visualization",
+  "version": "1.0.0",
+  "description": "SV visualization using jQuery and d3.js",
+  "main": "index.js",
+  "author": "jjdcabasolo",
+  "license": "WTFPL",
+  "private": true,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/jjdcabasolo/sv-visualization"
+  },
+  "dependencies": {
+    "express": "^4.16.2",
+    "fastareader": "^0.2.0",
+    "sqlite3": "^4.0.0"
+  },
+  "devDependencies": {
+    "nodemon": "^1.17.3"
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  }
+}
