+++ public/css/main.css
@@ -38,8 +38,10 @@
 
 #sv-visualization-area {
 	position: absolute;
-	margin-top: -120px;
+	margin-top: -100px;
+	overflow-x: scroll;
 	width: 100%;
+	height: 100%;
 }
 
 svg#sv-visualization-ins,
@@ -60,6 +62,14 @@ svg#sv-visualization-dup {
 
 .remove-padding-bottom {
 	padding-bottom: 0 !important;
+}
+
+.remove-margin {
+	margin: 0 !important;
+}
+
+.remove-padding {
+	padding: 0 !important;
 }
 
 #expand-brush-options {
