+++ public/css/themes/default/assets/fonts/icons.woff2
