+++ public/js/saveSvgAsPng.js
